+++ sources/db/schema.rb
@@ -10,10 +10,26 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20161205001443) do
+ActiveRecord::Schema.define(version: 20161205102623) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
+
+  create_table "calls", force: :cascade do |t|
+    t.string   "uuid"
+    t.string   "caller_number"
+    t.string   "caller_name"
+    t.integer  "user_id"
+    t.integer  "user_number_id"
+    t.integer  "company_number_id"
+    t.string   "record_url"
+    t.datetime "created_at",        null: false
+    t.datetime "updated_at",        null: false
+    t.index ["company_number_id"], name: "index_calls_on_company_number_id", using: :btree
+    t.index ["user_id"], name: "index_calls_on_user_id", using: :btree
+    t.index ["user_number_id"], name: "index_calls_on_user_number_id", using: :btree
+    t.index ["uuid"], name: "index_calls_on_uuid", unique: true, using: :btree
+  end
 
   create_table "company_numbers", force: :cascade do |t|
     t.string   "sip_endpoint"
@@ -42,5 +58,8 @@ ActiveRecord::Schema.define(version: 20161205001443) do
     t.datetime "updated_at", null: false
   end
 
+  add_foreign_key "calls", "company_numbers"
+  add_foreign_key "calls", "user_numbers"
+  add_foreign_key "calls", "users"
   add_foreign_key "user_numbers", "users"
 end