+++ sources/db/migrate/20161205122429_add_name_to_company_number.rb
@@ -1,0 +1,5 @@
+class AddNameToCompanyNumber < ActiveRecord::Migration[5.0]
+  def change
+    add_column :company_numbers, :name, :string
+  end
+end