+++ CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.0.3] - 2016-12-05
 ### Added
  - Plivo call authentication.
  - The call model to save call history.
  - Calls controller and a index page listing the call history.
  - A name column to UserNumber and CompanyNumber.
+ - Save calls information during call flow.
 
 ### Changed
  - Added a step in deployment section in the README.
@@ -33,5 +36,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - A .gitignore for rails app.
  - A Rails skeleton.
 
-[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.0.2...HEAD
+[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.0.3...HEAD
+[0.0.3]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.0.1...v0.0.2